Split generic key parts into the separate module
